[CHORE] : Unit tests for Model (new iteration with asset bundle)
