+++ SayTheirNamesTests/Model/JSONDataModelAPIs.xcassets/search.dataset/search.json
@@ -1,0 +1,1 @@
+{"data":{"results":[{"searchable":{"id":23,"full_name":"George Floyd","identifier":"george-floyd","date_of_incident":"2020-05-25","number_of_children":2,"age":46,"city":"Minnesota","country":"United States","biography":null,"context":"Officers responding to a \"forgery in progress\" detained Floyd, kneeling on his neck for several minutes. Floyd did not resist arrest and repeatedly told the officers \"please\" and \"I can't breathe\".[2] Bystanders urged the officers to let Floyd up as he went quiet and stopped moving.","outcome":null,"sharable_links":{"base":"https:\/\/www.saythiernames.io\/people\/george-floyd","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/people\/george-floyd","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/people\/george-floyd","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/people\/george-floyd"}},"title":"george-floyd","url":null,"type":"people"},{"searchable":{"id":37,"full_name":"Jamel Floyd","identifier":"jamel-floyd","date_of_incident":"2020-06-03","number_of_children":0,"age":35,"city":"New York - NY","country":"United States","biography":null,"context":"Jamel Floyd, an inmate at the Metropolitan Detention Center, died of a heart attack after correctional officers pepper-sprayed him in his cell.","outcome":null,"sharable_links":{"base":"https:\/\/www.saythiernames.io\/people\/jamel-floyd","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/people\/jamel-floyd","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/people\/jamel-floyd","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/people\/jamel-floyd"}},"title":"jamel-floyd","url":null,"type":"people"},{"searchable":{"id":3,"person_id":23,"type_id":1,"identifier":"donate-to-george-floyd","banner_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","title":"Donate to George Floyd","description":"Donate George Floyd by donating here","outcome":null,"outcome_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","link":"https:\/\/www.gofundme.com\/f\/georgefloyd","sharable_links":{"base":"https:\/\/www.saythiernames.io\/donations\/","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/donations\/","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/donations\/","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/donations\/"}},"title":"3","url":null,"type":"donation_links"},{"searchable":{"id":5,"person_id":37,"type_id":1,"identifier":"donate-to-jamel-floyd","banner_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","title":"Donate to Jamel Floyd","description":"Donate Jamel Floyd by donating here","outcome":null,"outcome_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","link":"https:\/\/www.gofundme.com\/f\/justice-for-jamel-floyd","sharable_links":{"base":"https:\/\/www.saythiernames.io\/donations\/","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/donations\/","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/donations\/","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/donations\/"}},"title":"5","url":null,"type":"donation_links"},{"searchable":{"id":1,"person_id":2,"type_id":1,"identifier":"petition-for-james-scurlock","banner_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","title":"Petition For James Scurlock","description":"Help bring justice to James Scurlock by signing this petition","outcome":null,"outcome_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","link":"https:\/\/www.change.org\/p\/demand-justice-for-james-scurlock-killed-by-omaha-racist-during-george-floyd-protest","sharable_links":{"base":"https:\/\/www.saythiernames.io\/petition\/","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/petition\/","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/petition\/","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/petition\/"}},"title":"1","url":null,"type":"petition_links"},{"searchable":{"id":10,"person_id":23,"type_id":1,"identifier":"petition-for-george-floyd","banner_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","title":"Petition For George Floyd","description":"Help bring justice to George Floyd by signing this petition","outcome":null,"outcome_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","link":"https:\/\/www.change.org\/p\/mayor-jacob-frey-justice-for-george-floyd","sharable_links":{"base":"https:\/\/www.saythiernames.io\/petition\/","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/petition\/","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/petition\/","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/petition\/"}},"title":"10","url":null,"type":"petition_links"},{"searchable":{"id":15,"person_id":37,"type_id":1,"identifier":"petition-for-jamel-floyd","banner_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","title":"Petition For Jamel Floyd","description":"Help bring justice to Jamel Floyd by signing this petition","outcome":null,"outcome_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","link":"https:\/\/www.change.org\/p\/justice-for-jamel-floyd","sharable_links":{"base":"https:\/\/www.saythiernames.io\/petition\/","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/petition\/","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/petition\/","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/petition\/"}},"title":"15","url":null,"type":"petition_links"},{"searchable":{"id":66,"person_id":null,"type_id":2,"identifier":"justiceforfloyd","banner_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/donations\/black-lives-matter-fund\/black-lives-matter.jpg","title":"#JusticeForFloyd","description":"Demand the officers who killed George Floyd are charged with murder.","outcome":null,"outcome_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/donations\/black-lives-matter-fund\/black-lives-matter-1.jpg","link":"https:\/\/act.colorofchange.org\/sign\/justiceforfloyd_george_floyd_minneapolis","sharable_links":{"base":"https:\/\/www.saythiernames.io\/petition\/","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/petition\/","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/petition\/","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/petition\/"}},"title":"66","url":null,"type":"petition_links"}]}}
