--- SayTheirNamesTests/Model/Model_DonationTests.swift
@@ -1,64 +1,0 @@
-//
-//  Model_DonationTests.swift
-//  SayTheirNames
-//
-//  Copyright (c) 2020 Say Their Names Team (https://github.com/Say-Their-Name)
-//
-//  Permission is hereby granted, free of charge, to any person obtaining a copy
-//  of this software and associated documentation files (the "Software"), to deal
-//  in the Software without restriction, including without limitation the rights
-//  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-//  copies of the Software, and to permit persons to whom the Software is
-//  furnished to do so, subject to the following conditions:
-//
-//  The above copyright notice and this permission notice shall be included in
-//  all copies or substantial portions of the Software.
-//
-//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-//  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-//  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-//  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-//  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-//  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-//  THE SOFTWARE.
-
-import XCTest
-@testable import Say_Their_Names
-
-class Model_DonationTests: XCTestCase {
-
-    let decoder = JSONDecoder()
-    let donationsString = #"""
-{"data":[{"id":1,"identifier":"donate-to-regis-korchinski-paquet","title":"Donate to Regis Korchinski-Paquet","description":"Donate Regis Korchinski-Paquet by donating here","outcome":null,"link":"https:\/\/uk.gofundme.com\/f\/justice-for-regis","outcome_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","banner_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","sharable_links":{"base":"https:\/\/www.saythiernames.io\/donations\/","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/donations\/","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/donations\/","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/donations\/"},"person":{"id":18,"full_name":"Regis Korchinski-Paquet","identifier":"regis-korchinski-paquet","date_of_incident":1590537600,"number_of_children":null,"age":29,"city":"Toronto","country":"Canada","their_story":"The death of Regis Korchinski-Paquet, an indigenous-black Canadian woman, occurred in Toronto, Ontario, Canada on May 27, 2020. Her mother, Claudette Beals-Clayton, called for police assistance after a domestic conflict with a request to take her daughter to the Centre for Addiction and Mental Health (CAMH). While her mother and brother were in the hallway of their apartment building, police entered the 24th floor apartment after Korchinski-Paquet, who went to use the bathroom. The family heard Korchinski-Paquet cry for help, followed by quiet, and then the officers confirming that she had fallen off the balcony. Korchinski-Paquet fell from the balcony to the ground below, and died at the scene. Her family accused the officers of pushing her off the balcony, which has led to a Special Investigations Unit investigation. Korchinski-Paquet's death has inspired a number of protests in Canada against police involvement in her death and other issues of anti-black and anti-indigenous racism. These protests occurred at the same time as the George Floyd protests.","outcome":null,"biography":null,"images":[{"id":18,"person_id":18,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/regis-korchinskipaquet\/regis-korchinskipaquet.jpg"}],"sharable_links":{"base":"https:\/\/www.saythiernames.io\/people\/regis-korchinski-paquet","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/people\/regis-korchinski-paquet","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/people\/regis-korchinski-paquet","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/people\/regis-korchinski-paquet"},"media":[{"id":18,"person_id":18,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/regis-korchinskipaquet\/regis-korchinskipaquet.jpg"}]},"type":{"id":1,"type":"Victims"}},{"id":2,"identifier":"donate-to-michael-powell","title":"Donate to Michael Powell","description":"Donate Michael Powell by donating here","outcome":null,"link":"https:\/\/mikeypowell-campaign.org.uk\/donations\/","outcome_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","banner_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","sharable_links":{"base":"https:\/\/www.saythiernames.io\/donations\/","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/donations\/","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/donations\/","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/donations\/"},"person":{"id":19,"full_name":"Michael Powell","identifier":"michael-powell","date_of_incident":1062892800,"number_of_children":null,"age":38,"city":"Birmingham","country":"United Kingdom","their_story":"Mikey Powell, 38, tragically died of asphyxiation while in custody after an arrest over a disturbance at his home in Lozells. During an arrest, the dad-of-three was hit by a police vehicle, sprayed with CS gas, restrained on the ground - while suffering psychosis - and bundled into the back of a police van. Around six minutes later, his body was placed on a mattress in a cell at Thornhill Road station in Handsworth where officers noticed he was not breathing. Efforts to resuscitate him were made, but he was pronounced dead at City Hospital shortly after.West Midlands Police paid \u00a3300,000 compensation to the family.","outcome":null,"biography":null,"images":[{"id":19,"person_id":19,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/michael-powell\/michael-powell.jpg"}],"sharable_links":{"base":"https:\/\/www.saythiernames.io\/people\/michael-powell","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/people\/michael-powell","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/people\/michael-powell","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/people\/michael-powell"},"media":[{"id":19,"person_id":19,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/michael-powell\/michael-powell.jpg"}]},"type":{"id":1,"type":"Victims"}},{"id":3,"identifier":"donate-to-george-floyd","title":"Donate to George Floyd","description":"Donate George Floyd by donating here","outcome":null,"link":"https:\/\/www.gofundme.com\/f\/georgefloyd","outcome_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","banner_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","sharable_links":{"base":"https:\/\/www.saythiernames.io\/donations\/","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/donations\/","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/donations\/","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/donations\/"},"person":{"id":23,"full_name":"George Floyd","identifier":"george-floyd","date_of_incident":1590364800,"number_of_children":2,"age":46,"city":"Minnesota","country":"United States","their_story":"Officers responding to a \"forgery in progress\" detained Floyd, kneeling on his neck for several minutes. Floyd did not resist arrest and repeatedly told the officers \"please\" and \"I can't breathe\".[2] Bystanders urged the officers to let Floyd up as he went quiet and stopped moving.","outcome":null,"biography":null,"images":[{"id":23,"person_id":23,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/george-floyd\/george-floyd.jpg"},{"id":24,"person_id":23,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/george-floyd\/george-floyd-1.jpg"}],"sharable_links":{"base":"https:\/\/www.saythiernames.io\/people\/george-floyd","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/people\/george-floyd","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/people\/george-floyd","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/people\/george-floyd"},"media":[{"id":23,"person_id":23,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/george-floyd\/george-floyd.jpg"},{"id":24,"person_id":23,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/george-floyd\/george-floyd-1.jpg"}]},"type":{"id":1,"type":"Victims"}},{"id":4,"identifier":"donate-to-cherry-groce","title":"Donate to Cherry Groce","description":"Donate Cherry Groce by donating here","outcome":null,"link":"https:\/\/weare336.org.uk\/organisations\/cherry-groce-foundation\/","outcome_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","banner_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","sharable_links":{"base":"https:\/\/www.saythiernames.io\/donations\/","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/donations\/","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/donations\/","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/donations\/"},"person":{"id":36,"full_name":"Cherry Groce","identifier":"cherry-groce","date_of_incident":496713600,"number_of_children":8,"age":null,"city":"London","country":"United Kingdom","their_story":"On the morning of Saturday 28 September 1985, a group of police officers raided Cherry Groce's house on Normandy Road, Brixton, including an armed CID officer, searching for suspected armed robber Michael Groce. Mrs Groce was in bed, whilst three of her six children were in residence when the police entered the house with force. During the raid, Mrs Groce was shot in the chest; after an extensive search, the police did not find Michael Groce. An ambulance was called, and by the time it had arrived a small crowd had gathered outside the house, her shooting sparked the 1985 Brixton riots. Cherry was left paralysed from the waist down and died of kidney failure in 2011. She had been more susceptible to a host of debilitating illnesses as a result of the shooting injury.","outcome":null,"biography":null,"images":[{"id":37,"person_id":36,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/cherry-groce\/cherry-groce.jpg"}],"sharable_links":{"base":"https:\/\/www.saythiernames.io\/people\/cherry-groce","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/people\/cherry-groce","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/people\/cherry-groce","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/people\/cherry-groce"},"media":[{"id":37,"person_id":36,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/cherry-groce\/cherry-groce.jpg"}]},"type":{"id":1,"type":"Victims"}},{"id":5,"identifier":"donate-to-jamel-floyd","title":"Donate to Jamel Floyd","description":"Donate Jamel Floyd by donating here","outcome":null,"link":"https:\/\/www.gofundme.com\/f\/justice-for-jamel-floyd","outcome_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","banner_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","sharable_links":{"base":"https:\/\/www.saythiernames.io\/donations\/","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/donations\/","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/donations\/","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/donations\/"},"person":{"id":37,"full_name":"Jamel Floyd","identifier":"jamel-floyd","date_of_incident":1591142400,"number_of_children":null,"age":35,"city":"New York - NY","country":"United States","their_story":"Jamel Floyd, an inmate at the Metropolitan Detention Center, died of a heart attack after correctional officers pepper-sprayed him in his cell.","outcome":null,"biography":null,"images":[{"id":38,"person_id":37,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/jamel-floyd\/jamel-floyd.jpg"}],"sharable_links":{"base":"https:\/\/www.saythiernames.io\/people\/jamel-floyd","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/people\/jamel-floyd","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/people\/jamel-floyd","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/people\/jamel-floyd"},"media":[{"id":38,"person_id":37,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/jamel-floyd\/jamel-floyd.jpg"}]},"type":{"id":1,"type":"Victims"}},{"id":6,"identifier":"donate-to-maurice-gordon","title":"Donate to Maurice Gordon","description":"Donate Maurice Gordon by donating here","outcome":null,"link":"https:\/\/www.gofundme.com\/f\/justice-for-maurice-gordon\/donate","outcome_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","banner_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","sharable_links":{"base":"https:\/\/www.saythiernames.io\/donations\/","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/donations\/","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/donations\/","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/donations\/"},"person":{"id":43,"full_name":"Maurice Gordon","identifier":"maurice-gordon","date_of_incident":1590192000,"number_of_children":null,"age":28,"city":"Poughkeepsie - NY","country":"United States","their_story":"On May 23, a routine traffic stop ended with a New Jersey State Police trooper shooting a black man to death on the side of the Garden State Parkway. Maurice Gordon, 28, of Poughkeepsie, New York, was unarmed and waiting with a white trooper for a tow truck to arrive because his car wouldn\u2019t restart, according to the attorney for Gordon\u2019s family.","outcome":null,"biography":null,"images":[{"id":44,"person_id":43,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/maurice-gordon\/maurice-gordon.jpg"}],"sharable_links":{"base":"https:\/\/www.saythiernames.io\/people\/maurice-gordon","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/people\/maurice-gordon","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/people\/maurice-gordon","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/people\/maurice-gordon"},"media":[{"id":44,"person_id":43,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/maurice-gordon\/maurice-gordon.jpg"}]},"type":{"id":1,"type":"Victims"}},{"id":7,"identifier":"donate-to-joyce-clarke","title":"Donate to Joyce Clarke","description":"Donate Joyce Clarke by donating here","outcome":null,"link":"https:\/\/www.gofundme.com\/f\/in-memory-of-joyce-clarke","outcome_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","banner_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","sharable_links":{"base":"https:\/\/www.saythiernames.io\/donations\/","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/donations\/","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/donations\/","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/donations\/"},"person":{"id":56,"full_name":"Joyce Clarke","identifier":"joyce-clarke","date_of_incident":1568678400,"number_of_children":1,"age":29,"city":"Geraldton","country":"Australia","their_story":"Joyce Clark was shot during a mental health check by police.","outcome":null,"biography":null,"images":[{"id":57,"person_id":56,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/joyce-clarke\/joyce-clarke.jpg"}],"sharable_links":{"base":"https:\/\/www.saythiernames.io\/people\/joyce-clarke","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/people\/joyce-clarke","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/people\/joyce-clarke","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/people\/joyce-clarke"},"media":[{"id":57,"person_id":56,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/joyce-clarke\/joyce-clarke.jpg"}]},"type":{"id":1,"type":"Victims"}},{"id":8,"identifier":"donate-to-trevor-belle","title":"Donate to Trevor Belle","description":"Donate Trevor Belle by donating here","outcome":null,"link":"https:\/\/uk.gofundme.com\/f\/trevor-belle","outcome_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","banner_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","sharable_links":{"base":"https:\/\/www.saythiernames.io\/donations\/","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/donations\/","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/donations\/","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/donations\/"},"person":{"id":58,"full_name":"Trevor Belle","identifier":"trevor-belle","date_of_incident":1584835200,"number_of_children":null,"age":61,"city":"London","country":"United Kingdom","their_story":"Trevor Belle, 61, died in the Royal London Hospital on April 18, having tested positive for Covid-19. Several weeks earlier, the east Londoner had been spat at by a passenger who owed him \u00a39. The new grandfather died three days after his 61st birthday","outcome":null,"biography":null,"images":[{"id":59,"person_id":58,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/trevor-belle\/trevor-belle.jpg"}],"sharable_links":{"base":"https:\/\/www.saythiernames.io\/people\/trevor-belle","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/people\/trevor-belle","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/people\/trevor-belle","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/people\/trevor-belle"},"media":[{"id":59,"person_id":58,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/trevor-belle\/trevor-belle.jpg"}]},"type":{"id":1,"type":"Victims"}},{"id":9,"identifier":"donate-to-tanya-day","title":"Donate to Tanya Day","description":"Donate Tanya Day by donating here","outcome":null,"link":"https:\/\/au.gofundme.com\/f\/day-family-fundraiser","outcome_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","banner_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","sharable_links":{"base":"https:\/\/www.saythiernames.io\/donations\/","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/donations\/","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/donations\/","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/donations\/"},"person":{"id":69,"full_name":"Tanya Day","identifier":"tanya-day","date_of_incident":1513900800,"number_of_children":1,"age":55,"city":"Castlemaine - VIC","country":"Australia","their_story":"Tanya Day sustained treatment from police for public drunkenness which they would not impose on non-Aboriginal people. The coroner found her death preventable had she not been arrested and taken into custody. The Day family says \"The Coroner found that once mum was in the cell, the police failed to properly check on her, treated her with complacency and breached her human rights to humane treatment and dignity.\"","outcome":null,"biography":null,"images":[{"id":70,"person_id":69,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/tanya-day\/tanya-day.png"}],"sharable_links":{"base":"https:\/\/www.saythiernames.io\/people\/tanya-day","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/people\/tanya-day","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/people\/tanya-day","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/people\/tanya-day"},"media":[{"id":70,"person_id":69,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/tanya-day\/tanya-day.png"}]},"type":{"id":1,"type":"Victims"}},{"id":10,"identifier":"donate-to-tony-mcdade","title":"Donate to Tony McDade","description":"Donate Tony McDade by donating here","outcome":null,"link":"https:\/\/www.gofundme.com\/f\/in-memory-of-tony-mcdade\/donate","outcome_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","banner_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","sharable_links":{"base":"https:\/\/www.saythiernames.io\/donations\/","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/donations\/","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/donations\/","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/donations\/"},"person":{"id":74,"full_name":"Tony McDade","identifier":"tony-mcdade","date_of_incident":1590537600,"number_of_children":null,"age":38,"city":"Tallahassee - FL","country":"United States","their_story":"Tony was identified as a suspect in a reported stabbing the morning of his death. Police said McDade fled the scene of the stabbing on foot before officers arrived, but they encountered McDade nearby shortly afterward. According to police, McDade reportedly had a handgun and \"made a move consistent with using the firearm\" against an officer, prompting the officer to shoot McDade.","outcome":null,"biography":null,"images":[{"id":75,"person_id":74,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/tony-mcdade\/tony-mcdade.jpg"}],"sharable_links":{"base":"https:\/\/www.saythiernames.io\/people\/tony-mcdade","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/people\/tony-mcdade","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/people\/tony-mcdade","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/people\/tony-mcdade"},"media":[{"id":75,"person_id":74,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/tony-mcdade\/tony-mcdade.jpg"}]},"type":{"id":1,"type":"Victims"}},{"id":11,"identifier":"donate-to-belly-mujinga","title":"Donate to Belly Mujinga","description":"Donate Belly Mujinga by donating here","outcome":null,"link":"https:\/\/www.gofundme.com\/f\/rip-belly-mujinga","outcome_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","banner_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","sharable_links":{"base":"https:\/\/www.saythiernames.io\/donations\/","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/donations\/","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/donations\/","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/donations\/"},"person":{"id":75,"full_name":"Belly Mujinga","identifier":"belly-mujinga","date_of_incident":1586044800,"number_of_children":null,"age":47,"city":"London","country":"United Kingdom","their_story":"Mujinga, 49, died on 5 April after contracting Covid-19. She fell ill with the coronavirus days after it was alleged a man who said he had Covid-19 spat and coughed at her and a colleague at Victoria station in London.","outcome":null,"biography":null,"images":[{"id":76,"person_id":75,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/belly-mujinga\/belly-mujinga.jpg"}],"sharable_links":{"base":"https:\/\/www.saythiernames.io\/people\/belly-mujinga","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/people\/belly-mujinga","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/people\/belly-mujinga","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/people\/belly-mujinga"},"media":[{"id":76,"person_id":75,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/belly-mujinga\/belly-mujinga.jpg"}]},"type":{"id":1,"type":"Victims"}},{"id":12,"identifier":"donate-to-breonna-taylor","title":"Donate to Breonna Taylor","description":"Donate Breonna Taylor by donating here","outcome":null,"link":"https:\/\/www.gofundme.com\/f\/9v4q2-justice-for-breonna-taylor","outcome_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","banner_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","sharable_links":{"base":"https:\/\/www.saythiernames.io\/donations\/","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/donations\/","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/donations\/","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/donations\/"},"person":{"id":77,"full_name":"Breonna Taylor","identifier":"breonna-taylor","date_of_incident":1584057600,"number_of_children":null,"age":26,"city":"Louisville - KT","country":"United States","their_story":"On March 13, 2020, Breonna Taylor, a 26-year-old African American woman, was fatally shot by Louisville Metro Police Department officers. ","outcome":null,"biography":null,"images":[{"id":78,"person_id":77,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/breonna-taylor\/breonna-taylor.jpg"}],"sharable_links":{"base":"https:\/\/www.saythiernames.io\/people\/breonna-taylor","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/people\/breonna-taylor","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/people\/breonna-taylor","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/people\/breonna-taylor"},"media":[{"id":78,"person_id":77,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/breonna-taylor\/breonna-taylor.jpg"}]},"type":{"id":1,"type":"Victims"}}],"links":{"first":"https:\/\/saytheirnames.dev\/api\/donations?page=1","last":"https:\/\/saytheirnames.dev\/api\/donations?page=2","prev":null,"next":"https:\/\/saytheirnames.dev\/api\/donations?page=2"},"meta":{"current_page":1,"from":1,"last_page":2,"path":"https:\/\/saytheirnames.dev\/api\/donations","per_page":12,"to":12,"total":20}}
-"""#
-    
-    override func setUpWithError() throws {
-        // Put setup code here. This method is called before the invocation of each test method in the class.
-        decoder.dateDecodingStrategy = .deferredToDate
-    }
-
-    override func tearDownWithError() throws {
-        // Put teardown code here. This method is called after the invocation of each test method in the class.
-    }
-
-    func testInit() throws {
-        // This is an example of a functional test case.
-        // Use XCTAssert and related functions to verify your tests produce the correct results.
-        
-        let json = Data(donationsString.utf8)
-        let donationsResponse = try decoder.decode(DonationsResponsePage.self, from: json)
-        
-        XCTAssertNil(donationsResponse.link.prev, "Previous link should be nil.")
-        XCTAssertNotNil(donationsResponse.link.next, "Next link should not be nil.")
-        
-        XCTAssertEqual(donationsResponse.all.count, 12, "There should be exactly 12 donations.")
-    }
-
-    func testPerformanceExample() throws {
-        // This is an example of a performance test case.
-        self.measure {
-            // Put the code you want to measure the time of here.
-        }
-    }
-
-}
