--- SayTheirNamesTests/Model/Model_PeopleTests.swift
@@ -1,68 +1,0 @@
-//
-//  Model_PeopleTests.swift
-//  SayTheirNames
-//
-//  Copyright (c) 2020 Say Their Names Team (https://github.com/Say-Their-Name)
-//
-//  Permission is hereby granted, free of charge, to any person obtaining a copy
-//  of this software and associated documentation files (the "Software"), to deal
-//  in the Software without restriction, including without limitation the rights
-//  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-//  copies of the Software, and to permit persons to whom the Software is
-//  furnished to do so, subject to the following conditions:
-//
-//  The above copyright notice and this permission notice shall be included in
-//  all copies or substantial portions of the Software.
-//
-//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-//  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-//  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-//  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-//  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-//  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-//  THE SOFTWARE.
-
-import XCTest
-@testable import Say_Their_Names
-
-class Model_PeopleTests: XCTestCase {
-
-    let decoder = JSONDecoder()
-    let peopleString = #"""
-{"data":[{"id":37,"full_name":"Jamel Floyd","identifier":"jamel-floyd","date_of_incident":1591142400,"number_of_children":null,"age":35,"city":"New York - NY","country":"United States","their_story":"Jamel Floyd, an inmate at the Metropolitan Detention Center, died of a heart attack after correctional officers pepper-sprayed him in his cell.","outcome":null,"biography":null,"images":[{"id":38,"person_id":37,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/jamel-floyd\/jamel-floyd.jpg"}],"sharable_links":{"base":"https:\/\/www.saythiernames.io\/people\/jamel-floyd","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/people\/jamel-floyd","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/people\/jamel-floyd","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/people\/jamel-floyd"},"media":[{"id":38,"person_id":37,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/jamel-floyd\/jamel-floyd.jpg"}]},{"id":99,"full_name":"David McAtee","identifier":"david-mcatee","date_of_incident":1590969600,"number_of_children":null,"age":53,"city":"Louisville - KT","country":"United States","their_story":"Louisville, Ky., resident David McAtee was shot and killed during a police response to a curfew violation","outcome":null,"biography":null,"images":[{"id":101,"person_id":99,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/david-mcatee\/david-mcatee.jpg"}],"sharable_links":{"base":"https:\/\/www.saythiernames.io\/people\/david-mcatee","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/people\/david-mcatee","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/people\/david-mcatee","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/people\/david-mcatee"},"media":[{"id":101,"person_id":99,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/david-mcatee\/david-mcatee.jpg"}]},{"id":2,"full_name":"James Scurlock","identifier":"james-scurlock","date_of_incident":1590796800,"number_of_children":null,"age":22,"city":"Omaha - NE","country":"United States","their_story":"Scurlock, a 22-year-old black man, was fatally shot by a white bar owner during a fight with several people one night in Omaha, Nebraska, amid protests related to the police killing of George Floyd. ","outcome":null,"biography":null,"images":[{"id":2,"person_id":2,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/james-scurlock\/james-scurlock.jpg"}],"sharable_links":{"base":"https:\/\/www.saythiernames.io\/people\/james-scurlock","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/people\/james-scurlock","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/people\/james-scurlock","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/people\/james-scurlock"},"media":[{"id":2,"person_id":2,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/james-scurlock\/james-scurlock.jpg"}]},{"id":18,"full_name":"Regis Korchinski-Paquet","identifier":"regis-korchinski-paquet","date_of_incident":1590537600,"number_of_children":null,"age":29,"city":"Toronto","country":"Canada","their_story":"The death of Regis Korchinski-Paquet, an indigenous-black Canadian woman, occurred in Toronto, Ontario, Canada on May 27, 2020. Her mother, Claudette Beals-Clayton, called for police assistance after a domestic conflict with a request to take her daughter to the Centre for Addiction and Mental Health (CAMH). While her mother and brother were in the hallway of their apartment building, police entered the 24th floor apartment after Korchinski-Paquet, who went to use the bathroom. The family heard Korchinski-Paquet cry for help, followed by quiet, and then the officers confirming that she had fallen off the balcony. Korchinski-Paquet fell from the balcony to the ground below, and died at the scene. Her family accused the officers of pushing her off the balcony, which has led to a Special Investigations Unit investigation. Korchinski-Paquet's death has inspired a number of protests in Canada against police involvement in her death and other issues of anti-black and anti-indigenous racism. These protests occurred at the same time as the George Floyd protests.","outcome":null,"biography":null,"images":[{"id":18,"person_id":18,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/regis-korchinskipaquet\/regis-korchinskipaquet.jpg"}],"sharable_links":{"base":"https:\/\/www.saythiernames.io\/people\/regis-korchinski-paquet","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/people\/regis-korchinski-paquet","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/people\/regis-korchinski-paquet","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/people\/regis-korchinski-paquet"},"media":[{"id":18,"person_id":18,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/regis-korchinskipaquet\/regis-korchinskipaquet.jpg"}]},{"id":74,"full_name":"Tony McDade","identifier":"tony-mcdade","date_of_incident":1590537600,"number_of_children":null,"age":38,"city":"Tallahassee - FL","country":"United States","their_story":"Tony was identified as a suspect in a reported stabbing the morning of his death. Police said McDade fled the scene of the stabbing on foot before officers arrived, but they encountered McDade nearby shortly afterward. According to police, McDade reportedly had a handgun and \"made a move consistent with using the firearm\" against an officer, prompting the officer to shoot McDade.","outcome":null,"biography":null,"images":[{"id":75,"person_id":74,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/tony-mcdade\/tony-mcdade.jpg"}],"sharable_links":{"base":"https:\/\/www.saythiernames.io\/people\/tony-mcdade","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/people\/tony-mcdade","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/people\/tony-mcdade","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/people\/tony-mcdade"},"media":[{"id":75,"person_id":74,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/tony-mcdade\/tony-mcdade.jpg"}]},{"id":23,"full_name":"George Floyd","identifier":"george-floyd","date_of_incident":1590364800,"number_of_children":2,"age":46,"city":"Minnesota","country":"United States","their_story":"Officers responding to a \"forgery in progress\" detained Floyd, kneeling on his neck for several minutes. Floyd did not resist arrest and repeatedly told the officers \"please\" and \"I can't breathe\".[2] Bystanders urged the officers to let Floyd up as he went quiet and stopped moving.","outcome":null,"biography":null,"images":[{"id":23,"person_id":23,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/george-floyd\/george-floyd.jpg"},{"id":24,"person_id":23,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/george-floyd\/george-floyd-1.jpg"}],"sharable_links":{"base":"https:\/\/www.saythiernames.io\/people\/george-floyd","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/people\/george-floyd","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/people\/george-floyd","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/people\/george-floyd"},"media":[{"id":23,"person_id":23,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/george-floyd\/george-floyd.jpg"},{"id":24,"person_id":23,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/george-floyd\/george-floyd-1.jpg"}]},{"id":43,"full_name":"Maurice Gordon","identifier":"maurice-gordon","date_of_incident":1590192000,"number_of_children":null,"age":28,"city":"Poughkeepsie - NY","country":"United States","their_story":"On May 23, a routine traffic stop ended with a New Jersey State Police trooper shooting a black man to death on the side of the Garden State Parkway. Maurice Gordon, 28, of Poughkeepsie, New York, was unarmed and waiting with a white trooper for a tow truck to arrive because his car wouldn\u2019t restart, according to the attorney for Gordon\u2019s family.","outcome":null,"biography":null,"images":[{"id":44,"person_id":43,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/maurice-gordon\/maurice-gordon.jpg"}],"sharable_links":{"base":"https:\/\/www.saythiernames.io\/people\/maurice-gordon","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/people\/maurice-gordon","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/people\/maurice-gordon","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/people\/maurice-gordon"},"media":[{"id":44,"person_id":43,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/maurice-gordon\/maurice-gordon.jpg"}]},{"id":28,"full_name":"Jo\u0101o Pedro","identifier":"joao-pedro","date_of_incident":1589760000,"number_of_children":null,"age":14,"city":"Rio de Janeiro","country":"Brazil","their_story":"On May 18, three police officers, supposedly pursuing suspects, entered a home in the Salgueiro favela where six unarmed cousins had gathered to play. They opened fire hitting 14-year-old Jo\u00e3o Pedro Matos Pinto in the back. ","outcome":null,"biography":null,"images":[{"id":29,"person_id":28,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/joao-pedro\/joao-pedro.jpg"}],"sharable_links":{"base":"https:\/\/www.saythiernames.io\/people\/joao-pedro","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/people\/joao-pedro","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/people\/joao-pedro","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/people\/joao-pedro"},"media":[{"id":29,"person_id":28,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/joao-pedro\/joao-pedro.jpg"}]},{"id":75,"full_name":"Belly Mujinga","identifier":"belly-mujinga","date_of_incident":1586044800,"number_of_children":null,"age":47,"city":"London","country":"United Kingdom","their_story":"Mujinga, 49, died on 5 April after contracting Covid-19. She fell ill with the coronavirus days after it was alleged a man who said he had Covid-19 spat and coughed at her and a colleague at Victoria station in London.","outcome":null,"biography":null,"images":[{"id":76,"person_id":75,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/belly-mujinga\/belly-mujinga.jpg"}],"sharable_links":{"base":"https:\/\/www.saythiernames.io\/people\/belly-mujinga","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/people\/belly-mujinga","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/people\/belly-mujinga","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/people\/belly-mujinga"},"media":[{"id":76,"person_id":75,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/belly-mujinga\/belly-mujinga.jpg"}]},{"id":46,"full_name":"Yassin Hussein Moyo","identifier":"yassin-hussein-moyo","date_of_incident":1585699200,"number_of_children":null,"age":13,"city":"Nairobi","country":"Kenya","their_story":"13-year-old Yassin Moyo who was shot dead by police officers in Kiamaiko during a curfew enforcement patrol by police officers believed to be from Huruma Police Station","outcome":null,"biography":null,"images":[{"id":47,"person_id":46,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/yassin-hussein-moyo\/yassinhussein-moyo.jpg"}],"sharable_links":{"base":"https:\/\/www.saythiernames.io\/people\/yassin-hussein-moyo","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/people\/yassin-hussein-moyo","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/people\/yassin-hussein-moyo","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/people\/yassin-hussein-moyo"},"media":[{"id":47,"person_id":46,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/yassin-hussein-moyo\/yassinhussein-moyo.jpg"}]},{"id":58,"full_name":"Trevor Belle","identifier":"trevor-belle","date_of_incident":1584835200,"number_of_children":null,"age":61,"city":"London","country":"United Kingdom","their_story":"Trevor Belle, 61, died in the Royal London Hospital on April 18, having tested positive for Covid-19. Several weeks earlier, the east Londoner had been spat at by a passenger who owed him \u00a39. The new grandfather died three days after his 61st birthday","outcome":null,"biography":null,"images":[{"id":59,"person_id":58,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/trevor-belle\/trevor-belle.jpg"}],"sharable_links":{"base":"https:\/\/www.saythiernames.io\/people\/trevor-belle","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/people\/trevor-belle","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/people\/trevor-belle","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/people\/trevor-belle"},"media":[{"id":59,"person_id":58,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/trevor-belle\/trevor-belle.jpg"}]},{"id":77,"full_name":"Breonna Taylor","identifier":"breonna-taylor","date_of_incident":1584057600,"number_of_children":null,"age":26,"city":"Louisville - KT","country":"United States","their_story":"On March 13, 2020, Breonna Taylor, a 26-year-old African American woman, was fatally shot by Louisville Metro Police Department officers. ","outcome":null,"biography":null,"images":[{"id":78,"person_id":77,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/breonna-taylor\/breonna-taylor.jpg"}],"sharable_links":{"base":"https:\/\/www.saythiernames.io\/people\/breonna-taylor","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/people\/breonna-taylor","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/people\/breonna-taylor","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/people\/breonna-taylor"},"media":[{"id":78,"person_id":77,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/breonna-taylor\/breonna-taylor.jpg"}]}],"links":{"first":"https:\/\/saytheirnames.dev\/api\/people?page=1","last":"https:\/\/saytheirnames.dev\/api\/people?page=10","prev":null,"next":"https:\/\/saytheirnames.dev\/api\/people?page=2"},"meta":{"current_page":1,"from":1,"last_page":10,"path":"https:\/\/saytheirnames.dev\/api\/people","per_page":12,"to":12,"total":117}}
-"""#
-    
-    override func setUpWithError() throws {
-        // Put setup code here. This method is called before the invocation of each test method in the class.
-        
-        let dateFormatter = DateFormatter()
-        dateFormatter.locale = Locale(identifier: "en_US")
-        dateFormatter.setLocalizedDateFormatFromTemplate("yyyy-MM-d")
-        decoder.dateDecodingStrategy = .formatted(dateFormatter)
-    }
-
-    override func tearDownWithError() throws {
-        // Put teardown code here. This method is called after the invocation of each test method in the class.
-    }
-
-    func testInit() throws {
-        // This is an example of a functional test case.
-        // Use XCTAssert and related functions to verify your tests produce the correct results.
-        
-        let json = Data(peopleString.utf8)
-        let personsResponse = try decoder.decode(PersonsResponsePage.self, from: json)
-        
-        XCTAssertNil(personsResponse.link.prev, "Previous link should be nil.")
-        XCTAssertNotNil(personsResponse.link.next, "Next link should not be nil.")
-        
-        XCTAssertEqual(personsResponse.all.count, 12, "There should be exactly 12 persons.")
-    }
-
-    func testPerformanceExample() throws {
-        // This is an example of a performance test case.
-        self.measure {
-            // Put the code you want to measure the time of here.
-        }
-    }
-
-}
