--- SayTheirNamesTests/Model/Model_PetitionsTests.swift
@@ -1,62 +1,0 @@
-//
-//  Model_PetitionsTests.swift
-//  SayTheirNames
-//
-//  Copyright (c) 2020 Say Their Names Team (https://github.com/Say-Their-Name)
-//
-//  Permission is hereby granted, free of charge, to any person obtaining a copy
-//  of this software and associated documentation files (the "Software"), to deal
-//  in the Software without restriction, including without limitation the rights
-//  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-//  copies of the Software, and to permit persons to whom the Software is
-//  furnished to do so, subject to the following conditions:
-//
-//  The above copyright notice and this permission notice shall be included in
-//  all copies or substantial portions of the Software.
-//
-//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-//  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-//  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-//  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-//  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-//  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-//  THE SOFTWARE.
-
-import XCTest
-@testable import Say_Their_Names
-
-class Model_PetitionsTests: XCTestCase {
-    
-    let decoder = JSONDecoder()
-    let petitionsString = #"""
-{"data":[{"id":1,"identifier":"petition-for-james-scurlock","title":"Petition For James Scurlock","description":"Help bring justice to James Scurlock by signing this petition","outcome":null,"link":"https:\/\/www.change.org\/p\/demand-justice-for-james-scurlock-killed-by-omaha-racist-during-george-floyd-protest","outcome_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","banner_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","sharable_links":{"base":"https:\/\/www.saythiernames.io\/petition\/","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/petition\/","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/petition\/","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/petition\/"},"person":{"id":2,"full_name":"James Scurlock","identifier":"james-scurlock","date_of_incident":1590796800,"number_of_children":null,"age":22,"city":"Omaha - NE","country":"United States","their_story":"Scurlock, a 22-year-old black man, was fatally shot by a white bar owner during a fight with several people one night in Omaha, Nebraska, amid protests related to the police killing of George Floyd. ","outcome":null,"biography":null,"images":[{"id":2,"person_id":2,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/james-scurlock\/james-scurlock.jpg"}],"sharable_links":{"base":"https:\/\/www.saythiernames.io\/people\/james-scurlock","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/people\/james-scurlock","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/people\/james-scurlock","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/people\/james-scurlock"},"media":[{"id":2,"person_id":2,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/james-scurlock\/james-scurlock.jpg"}]},"type":{"id":1,"type":"Victims"}},{"id":2,"identifier":"petition-for-olaseni-lewis","title":"Petition For Olaseni Lewis","description":"Help bring justice to Olaseni Lewis by signing this petition","outcome":null,"link":"http:\/\/www.justiceforseni.com\/","outcome_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","banner_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","sharable_links":{"base":"https:\/\/www.saythiernames.io\/petition\/","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/petition\/","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/petition\/","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/petition\/"},"person":{"id":7,"full_name":"Olaseni Lewis","identifier":"olaseni-lewis","date_of_incident":1283212800,"number_of_children":null,"age":23,"city":"London","country":"United Kingdom","their_story":"IT graduate Mr Lewis - who was also known as Seni - never regained consciousness after he was restrained for two prolonged periods of 10 and 20 minutes, at Bethlem Royal Hospital, Beckenham.","outcome":null,"biography":null,"images":[{"id":7,"person_id":7,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/olaseni-lewis\/olaseni-lewis.jpg"}],"sharable_links":{"base":"https:\/\/www.saythiernames.io\/people\/olaseni-lewis","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/people\/olaseni-lewis","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/people\/olaseni-lewis","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/people\/olaseni-lewis"},"media":[{"id":7,"person_id":7,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/olaseni-lewis\/olaseni-lewis.jpg"}]},"type":{"id":1,"type":"Victims"}},{"id":3,"identifier":"petition-for-jamar-clark","title":"Petition For Jamar Clark","description":"Help bring justice to Jamar Clark by signing this petition","outcome":null,"link":"https:\/\/campaigns.organizefor.org\/petitions\/no-grand-jury-justice4jamar-1","outcome_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","banner_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","sharable_links":{"base":"https:\/\/www.saythiernames.io\/petition\/","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/petition\/","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/petition\/","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/petition\/"},"person":{"id":9,"full_name":"Jamar Clark","identifier":"jamar-clark","date_of_incident":1447545600,"number_of_children":null,"age":24,"city":"Minneapolis - MN","country":"United States","their_story":"On November 15, 2015, Jamar Clark, a 24-year-old African-American man, was shot by Minneapolis Police in Minneapolis, Minnesota. Two police officers, Mark Ringgenberg and Dustin Schwarze, were involved in the shooting and were subsequently placed on paid administrative leave. The night after the shooting, Clark died at the Hennepin County Medical Center after being taken off life support.","outcome":null,"biography":null,"images":[{"id":9,"person_id":9,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/jamar-clark\/jamar-clark.jpg"}],"sharable_links":{"base":"https:\/\/www.saythiernames.io\/people\/jamar-clark","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/people\/jamar-clark","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/people\/jamar-clark","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/people\/jamar-clark"},"media":[{"id":9,"person_id":9,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/jamar-clark\/jamar-clark.jpg"}]},"type":{"id":1,"type":"Victims"}},{"id":4,"identifier":"petition-for-ramarley-graham","title":"Petition For Ramarley Graham","description":"Help bring justice to Ramarley Graham by signing this petition","outcome":null,"link":"https:\/\/www.justicecommittee.org\/riseup4ramarley","outcome_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","banner_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","sharable_links":{"base":"https:\/\/www.saythiernames.io\/petition\/","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/petition\/","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/petition\/","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/petition\/"},"person":{"id":10,"full_name":"Ramarley Graham","identifier":"ramarley-graham","date_of_incident":1328140800,"number_of_children":null,"age":18,"city":"New York - NY","country":"United States","their_story":"Graham died on February 2, 2012, at the age of 18 after being shot by a New York Police Department officer in his grandmother\u2019s bathroom. Ramarley Graham was spotted adjusting the waistband of his pants during a NYPD narcotics squad surveillance of a bodega near his home. Officers assumed Graham was concealing a gun. When plainclothes officers approached Graham, they reported that he ran to the home of Patricia Hartley, his grandmother, a few blocks away. Surveillance footage from outside the front entrance to Hartley\u2019s building, however, shows Graham calmly walking up to the door and entering without urgency. The same surveillance footage shows two officers running up to the building with their guns drawn just moments after Graham enters. As the two officers try to force their way through the locked door, other NYPD officers gained entrance through the rear of the building and opened the front door. One officer ran upstairs to Hartley\u2019s apartment and broke down the door without a search warrant or announcing that he was a police officer. Plainclothes Officer Richard Haste then ran into the apartment and discovered Graham in the bathroom attempting to flush a bag of marijuana down the toilet. Without any orders to comply, Graham was shot once in the chest in front of his little brother, Chinoor Campbell, and grandmother. Haste later reported that Graham reached into his waistband for a gun but no weapon was found in the Hartley apartment. In 2012, a grand jury voted to indict Haste, but the indictment was tossed out due to a prosecutorial mistake. A second grand jury declined to indict Haste, allowing him to walk free. After further investigation during a departmental trial in March 2017, Haste was found guilty on all charges and turned in his badge and gun before Deputy Commissioner of Trials, Rosemarie Maldonado, asked Police Commissioner, James O\u2019Neill, to fire him","outcome":null,"biography":null,"images":[{"id":10,"person_id":10,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/ramarley-graham\/ramarley-graham.jpg"}],"sharable_links":{"base":"https:\/\/www.saythiernames.io\/people\/ramarley-graham","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/people\/ramarley-graham","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/people\/ramarley-graham","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/people\/ramarley-graham"},"media":[{"id":10,"person_id":10,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/ramarley-graham\/ramarley-graham.jpg"}]},"type":{"id":1,"type":"Victims"}},{"id":5,"identifier":"petition-for-edson-da-costa","title":"Petition For Edson Da Costa","description":"Help bring justice to Edson Da Costa by signing this petition","outcome":null,"link":"https:\/\/www.change.org\/p\/metropolitan-police-commissioner-suspend-police-officers-under-investigation-for-deaths-of-edson-da-costa-rashan-charles","outcome_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","banner_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","sharable_links":{"base":"https:\/\/www.saythiernames.io\/petition\/","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/petition\/","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/petition\/","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/petition\/"},"person":{"id":11,"full_name":"Edson Da Costa","identifier":"edson-da-costa","date_of_incident":1498003200,"number_of_children":2,"age":25,"city":"London","country":"United Kingdom","their_story":"Edson Da Costa died after being detained by London police using force","outcome":null,"biography":null,"images":[{"id":11,"person_id":11,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/edson-dacosta\/edson-dacosta.jpg"}],"sharable_links":{"base":"https:\/\/www.saythiernames.io\/people\/edson-da-costa","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/people\/edson-da-costa","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/people\/edson-da-costa","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/people\/edson-da-costa"},"media":[{"id":11,"person_id":11,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/edson-dacosta\/edson-dacosta.jpg"}]},"type":{"id":1,"type":"Victims"}},{"id":6,"identifier":"petition-for-kimani-gray","title":"Petition For Kimani Gray","description":"Help bring justice to Kimani Gray by signing this petition","outcome":null,"link":"https:\/\/www.thepetitionsite.com\/en-gb\/597\/528\/146\/demand-justice-for-kimani-gray\/","outcome_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","banner_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","sharable_links":{"base":"https:\/\/www.saythiernames.io\/petition\/","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/petition\/","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/petition\/","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/petition\/"},"person":{"id":12,"full_name":"Kimani Gray","identifier":"kimani-gray","date_of_incident":1362787200,"number_of_children":null,"age":16,"city":"New York - NY","country":"United States","their_story":"Kimani Gray was 16 when he was shot 7 times and killed by undercover police. The complete story is unclear. Officers say Gray had a .38 and pointed it at them, some eyewitnesses said Gray did not have a gun. Officers said they yelled \u201cdon\u2019t move,\u201d but others said they did not identify themselves as NYPD and Gray may have thought he was being robbed. The two police involved in the shooting, Sgt. Mourad Mourad and Officer Jovaniel Cordova, did not face any charges as a result of the shooting. ","outcome":null,"biography":null,"images":[{"id":12,"person_id":12,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/kimani-gray\/kimani-gray.jpg"}],"sharable_links":{"base":"https:\/\/www.saythiernames.io\/people\/kimani-gray","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/people\/kimani-gray","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/people\/kimani-gray","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/people\/kimani-gray"},"media":[{"id":12,"person_id":12,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/kimani-gray\/kimani-gray.jpg"}]},"type":{"id":1,"type":"Victims"}},{"id":7,"identifier":"petition-for-julian-cole","title":"Petition For Julian Cole","description":"Help bring justice to Julian Cole by signing this petition","outcome":null,"link":"https:\/\/colefamilytruthcampaign.wordpress.com\/","outcome_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","banner_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","sharable_links":{"base":"https:\/\/www.saythiernames.io\/petition\/","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/petition\/","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/petition\/","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/petition\/"},"person":{"id":15,"full_name":"Julian Cole","identifier":"julian-cole","date_of_incident":1367798400,"number_of_children":null,"age":19,"city":"Bedford","country":"United Kingdom","their_story":"Witnesses saw Julian dragged unconscious across the road by officers in the direction of a police van. By this stage his neck had been broken; he had suffered a serious spinal injury and was unresponsive.","outcome":null,"biography":null,"images":[{"id":15,"person_id":15,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/julian-cole\/julian-cole.jpg"}],"sharable_links":{"base":"https:\/\/www.saythiernames.io\/people\/julian-cole","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/people\/julian-cole","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/people\/julian-cole","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/people\/julian-cole"},"media":[{"id":15,"person_id":15,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/julian-cole\/julian-cole.jpg"}]},"type":{"id":1,"type":"Victims"}},{"id":8,"identifier":"petition-for-michelle-shirley","title":"Petition For Michelle Shirley","description":"Help bring justice to Michelle Shirley by signing this petition","outcome":null,"link":"https:\/\/www.nomoremartyrs.org\/blog\/2017\/3\/5\/sayhername-michelle-shirley","outcome_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","banner_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","sharable_links":{"base":"https:\/\/www.saythiernames.io\/petition\/","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/petition\/","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/petition\/","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/petition\/"},"person":{"id":17,"full_name":"Michelle Shirley","identifier":"michelle-shirley","date_of_incident":1477872000,"number_of_children":1,"age":39,"city":"Torrance - CA","country":"United States","their_story":"Michelle, 39, was shot multile times and killed by police in Torrance who had been chasing Shirley and had cornered her at a gas station. Shirley suffered from bipolar disorder.","outcome":null,"biography":null,"images":[{"id":17,"person_id":17,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/michelle-shirley\/michelle-shirley.jpg"}],"sharable_links":{"base":"https:\/\/www.saythiernames.io\/people\/michelle-shirley","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/people\/michelle-shirley","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/people\/michelle-shirley","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/people\/michelle-shirley"},"media":[{"id":17,"person_id":17,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/michelle-shirley\/michelle-shirley.jpg"}]},"type":{"id":1,"type":"Victims"}},{"id":9,"identifier":"petition-for-regis-korchinski-paquet","title":"Petition For Regis Korchinski-Paquet","description":"Help bring justice to Regis Korchinski-Paquet by signing this petition","outcome":null,"link":"https:\/\/www.change.org\/p\/justice-for-regis-korchinski-paquet","outcome_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","banner_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","sharable_links":{"base":"https:\/\/www.saythiernames.io\/petition\/","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/petition\/","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/petition\/","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/petition\/"},"person":{"id":18,"full_name":"Regis Korchinski-Paquet","identifier":"regis-korchinski-paquet","date_of_incident":1590537600,"number_of_children":null,"age":29,"city":"Toronto","country":"Canada","their_story":"The death of Regis Korchinski-Paquet, an indigenous-black Canadian woman, occurred in Toronto, Ontario, Canada on May 27, 2020. Her mother, Claudette Beals-Clayton, called for police assistance after a domestic conflict with a request to take her daughter to the Centre for Addiction and Mental Health (CAMH). While her mother and brother were in the hallway of their apartment building, police entered the 24th floor apartment after Korchinski-Paquet, who went to use the bathroom. The family heard Korchinski-Paquet cry for help, followed by quiet, and then the officers confirming that she had fallen off the balcony. Korchinski-Paquet fell from the balcony to the ground below, and died at the scene. Her family accused the officers of pushing her off the balcony, which has led to a Special Investigations Unit investigation. Korchinski-Paquet's death has inspired a number of protests in Canada against police involvement in her death and other issues of anti-black and anti-indigenous racism. These protests occurred at the same time as the George Floyd protests.","outcome":null,"biography":null,"images":[{"id":18,"person_id":18,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/regis-korchinskipaquet\/regis-korchinskipaquet.jpg"}],"sharable_links":{"base":"https:\/\/www.saythiernames.io\/people\/regis-korchinski-paquet","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/people\/regis-korchinski-paquet","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/people\/regis-korchinski-paquet","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/people\/regis-korchinski-paquet"},"media":[{"id":18,"person_id":18,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/regis-korchinskipaquet\/regis-korchinskipaquet.jpg"}]},"type":{"id":1,"type":"Victims"}},{"id":10,"identifier":"petition-for-george-floyd","title":"Petition For George Floyd","description":"Help bring justice to George Floyd by signing this petition","outcome":null,"link":"https:\/\/www.change.org\/p\/mayor-jacob-frey-justice-for-george-floyd","outcome_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","banner_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","sharable_links":{"base":"https:\/\/www.saythiernames.io\/petition\/","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/petition\/","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/petition\/","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/petition\/"},"person":{"id":23,"full_name":"George Floyd","identifier":"george-floyd","date_of_incident":1590364800,"number_of_children":2,"age":46,"city":"Minnesota","country":"United States","their_story":"Officers responding to a \"forgery in progress\" detained Floyd, kneeling on his neck for several minutes. Floyd did not resist arrest and repeatedly told the officers \"please\" and \"I can't breathe\".[2] Bystanders urged the officers to let Floyd up as he went quiet and stopped moving.","outcome":null,"biography":null,"images":[{"id":23,"person_id":23,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/george-floyd\/george-floyd.jpg"},{"id":24,"person_id":23,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/george-floyd\/george-floyd-1.jpg"}],"sharable_links":{"base":"https:\/\/www.saythiernames.io\/people\/george-floyd","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/people\/george-floyd","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/people\/george-floyd","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/people\/george-floyd"},"media":[{"id":23,"person_id":23,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/george-floyd\/george-floyd.jpg"},{"id":24,"person_id":23,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/george-floyd\/george-floyd-1.jpg"}]},"type":{"id":1,"type":"Victims"}},{"id":11,"identifier":"petition-for-joao-pedro","title":"Petition For Jo\u0101o Pedro","description":"Help bring justice to Jo\u0101o Pedro by signing this petition","outcome":null,"link":"https:\/\/www.change.org\/p\/prefeitura-do-rio-de-janeiro-justice-for-jo%C4%81o-pedro","outcome_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","banner_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","sharable_links":{"base":"https:\/\/www.saythiernames.io\/petition\/","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/petition\/","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/petition\/","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/petition\/"},"person":{"id":28,"full_name":"Jo\u0101o Pedro","identifier":"joao-pedro","date_of_incident":1589760000,"number_of_children":null,"age":14,"city":"Rio de Janeiro","country":"Brazil","their_story":"On May 18, three police officers, supposedly pursuing suspects, entered a home in the Salgueiro favela where six unarmed cousins had gathered to play. They opened fire hitting 14-year-old Jo\u00e3o Pedro Matos Pinto in the back. ","outcome":null,"biography":null,"images":[{"id":29,"person_id":28,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/joao-pedro\/joao-pedro.jpg"}],"sharable_links":{"base":"https:\/\/www.saythiernames.io\/people\/joao-pedro","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/people\/joao-pedro","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/people\/joao-pedro","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/people\/joao-pedro"},"media":[{"id":29,"person_id":28,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/joao-pedro\/joao-pedro.jpg"}]},"type":{"id":1,"type":"Victims"}},{"id":12,"identifier":"petition-for-jermaine-baker","title":"Petition For Jermaine Baker","description":"Help bring justice to Jermaine Baker by signing this petition","outcome":null,"link":"https:\/\/petition.parliament.uk\/archived\/petitions\/163599","outcome_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","banner_img_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/petition.png","sharable_links":{"base":"https:\/\/www.saythiernames.io\/petition\/","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/petition\/","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/petition\/","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/petition\/"},"person":{"id":30,"full_name":"Jermaine Baker","identifier":"jermaine-baker","date_of_incident":1449792000,"number_of_children":2,"age":28,"city":"London","country":"United Kingdom","their_story":"Jermaine Baker, aged 28, a father-of-two from Tottenham, north London, was killed by a single shot to the neck from a police officer","outcome":null,"biography":null,"images":[{"id":31,"person_id":30,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/jermaine-baker\/jermaine-baker.jpg"}],"sharable_links":{"base":"https:\/\/www.saythiernames.io\/people\/jermaine-baker","facebook":"https:\/\/www.facebook.com\/sharer\/sharer.php?u=https:\/\/www.saythiernames.io\/people\/jermaine-baker","twitter":"https:\/\/twitter.com\/home?status=https:\/\/www.saythiernames.io\/people\/jermaine-baker","whatsapp":"https:\/\/wa.me\/?text=https:\/\/www.saythiernames.io\/people\/jermaine-baker"},"media":[{"id":31,"person_id":30,"image_url":"https:\/\/say-their-names.fra1.cdn.digitaloceanspaces.com\/jermaine-baker\/jermaine-baker.jpg"}]},"type":{"id":1,"type":"Victims"}}],"links":{"first":"https:\/\/saytheirnames.dev\/api\/petitions?page=1","last":"https:\/\/saytheirnames.dev\/api\/petitions?page=6","prev":null,"next":"https:\/\/saytheirnames.dev\/api\/petitions?page=2"},"meta":{"current_page":1,"from":1,"last_page":6,"path":"https:\/\/saytheirnames.dev\/api\/petitions","per_page":12,"to":12,"total":66}}
-"""#
-    override func setUpWithError() throws {
-        // Put setup code here. This method is called before the invocation of each test method in the class.
-    }
-
-    override func tearDownWithError() throws {
-        // Put teardown code here. This method is called after the invocation of each test method in the class.
-    }
-
-    func testInit() throws {
-        // This is an example of a functional test case.
-        // Use XCTAssert and related functions to verify your tests produce the correct results.
-        
-        let json = Data(petitionsString.utf8)
-        let petitionsResponse = try decoder.decode(PetitionsResponsePage.self, from: json)
-        
-        XCTAssertNil(petitionsResponse.link.prev, "Previous link should be nil.")
-        XCTAssertNotNil(petitionsResponse.link.next, "Next link should not be nil.")
-        
-        XCTAssertEqual(petitionsResponse.all.count, 12, "There should be exactly 12 petitions.")
-    }
-
-    func testPerformanceExample() throws {
-        // This is an example of a performance test case.
-        self.measure {
-            // Put the code you want to measure the time of here.
-        }
-    }
-
-}
